Remove dependency on adaptive content module, rewrite config
This changes the slice config to be grouped by template, rather than grouped
by setting name, which is a bit more natural and easier to read/manage:

    MySlice:
      templates:
        QuoteSlice:
          fields:
            Title:
              label: Title of the quote
              help: eg. "The great epic."

Rather than:

    MySlice:
      fieldLabels:
        QuoteSlice:
          Title: Title of the quote

      fieldHelp:
        QuoteSlice:
          Title: eg. "The great epic."

The hard dependency on silverstripe-adaptivecontent has also been removed,
which was a hang-on from this module originally being written into a
project and later extracted quickly. This commit is the start of aligning
this module with how we're currently using it. The original design of the
slices module was quite different to how most of our projects need to work,
so most projects use a bunch of traits and extensions (eg.
ClassBySecondaryIdentifierTrait from adaptive content).