Make ClassBySecondaryIdentifierTrait a default part of Slice
Making template names map to different slice data objects used to require
a trait from silverstripe-adaptivecontent. This is pretty key to how we
use slices, so it's now built in.

This introduces a requirement to add `getBaseSliceClass` to subclasses
of Slice that will be created directly. The functionality was previously
in a trait where __CLASS__ returns the class name of the class the trait
is applied to, effectively making this happen automatically.