Fix template includes not working in dataobject preview renders in CMS
Template includes work fine in the slice preview iframes in 0.x of this
module, but they're not working any more. I can't find the change that's
caused this to start being a problem (it doesn't appear to be SilverStripe
version), but this commit fixes the problem. Without the fix, template
rendering was triggering errors like:

  None of the following templates could be found (no theme in use): LibraryItemPopup_hb.ss

This also fixes the SliceWrapper.ss template not being pulled from the theme.