Prevent missing templates breaking the CMS when developing
Often when developing, the definition for a slice will be created before
its template. If an instance of the slice is then created, an exception
due to the missing template is thrown, which is mildly annoying. This change
allows creating a slice in the CMS before the template is created.